+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Factories for app development
+
 ## [0.2.0] - 2016-07-29
 ### Added
 - Add ability to create sender